Add Speld.nl (Dutch parody news site) (#1151)
* Add Speld.nl (Dutch parody news site)

Speld.nl is a popular parody new site in The Netherlands, not unlike The Onion.

The paywall is a max-number-of-visits type. The visits are tracked in a cookie, which this PR selectively removes to not trigger the cookie consent forms each time. We also remove a popup that says how many visits you've got left.

* typo in removeDOMElement

* Fix Speld.nl support